Refactor go-mkopensource to be able to reuse error handling functions in Python and Js code
